Reorganize test directory structure
Move tests from src/Bundle/test(s)/ to top-level tests/ directory
with proper PSR-4 namespace Sylius\Bundle\MailerBundle\Tests\

- tests/Application/ - test kernel and config
- tests/Unit/ - unit tests
- tests/Functional/ - functional tests
- tests/Integration/ - integration tests
