+++ composer.json
@@ -67,15 +67,12 @@
         "psr-4": {
             "Sylius\\Bundle\\MailerBundle\\": "src/Bundle/",
             "Sylius\\Component\\Mailer\\": "src/Component/"
-        },
-        "exclude-from-classmap": [
-            "src/Bundle/test/"
-        ]
+        }
     },
     "autoload-dev": {
         "psr-4": {
-            "Sylius\\Bundle\\MailerBundle\\Tests\\": ["src/Bundle/test/", "tests/"],
-            "App\\": "src/Bundle/test/src/"
+            "Sylius\\Bundle\\MailerBundle\\Tests\\": "tests/",
+            "App\\": "tests/Application/"
         }
     },
     "scripts": {
