+++ phpunit.xml.dist
@@ -3,6 +3,7 @@
 <phpunit xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:noNamespaceSchemaLocation="http://schema.phpunit.de/10.5/phpunit.xsd"
          colors="true"
+         bootstrap="tests/Application/config/bootstrap.php"
          defaultTestSuite="SyliusMailerBundle Test Suite"
 >
     <php>
@@ -11,12 +12,12 @@
 
     <testsuites>
         <testsuite name="SyliusMailerBundle Test Suite">
-            <directory>./src/Bundle/tests/</directory>
             <directory>./tests/</directory>
-            <exclude>./src/Bundle/tests/Integration/Cli/</exclude>
+            <exclude>./tests/Integration/</exclude>
+            <exclude>./tests/Application/</exclude>
         </testsuite>
         <testsuite name="SyliusMailerBundle Integration Test Suite">
-            <directory>./src/Bundle/tests/Integration/Cli/</directory>
+            <directory>./tests/Integration/</directory>
         </testsuite>
     </testsuites>
 </phpunit>
