--- src/Bundle/test/composer.json
@@ -1,3 +1,0 @@
-{
-    "name": "example/test-application"
-}
