+++ tests/Application/Kernel.php
@@ -19,4 +19,9 @@ use Symfony\Component\HttpKernel\Kernel as BaseKernel;
 class Kernel extends BaseKernel
 {
     use MicroKernelTrait;
+
+    public function getProjectDir(): string
+    {
+        return __DIR__;
+    }
 }
