+++ tests/Application/Modifier/EmailSenderNameModifier.php
