+++ tests/Application/config/bootstrap.php
@@ -13,7 +13,7 @@ declare(strict_types=1);
 
 use Symfony\Component\Dotenv\Dotenv;
 
-require dirname(__DIR__) . '/../../../vendor/autoload.php';
+require dirname(__DIR__, 3) . '/vendor/autoload.php';
 
 // Load cached env vars if the .env.local.php file exists
 // Run "composer dump-env prod" to create it (requires symfony/flex >=1.2)
