+++ tests/Application/config/packages/test_with_no_mailers/sylius_mailer.yaml
