+++ tests/Application/config/packages/test_with_translations/mailer.yaml
