+++ tests/Application/config/packages/test_with_translations/sylius_mailer.yaml
