+++ tests/Functional/DefaultSenderTest.php
@@ -11,10 +11,10 @@
 
 declare(strict_types=1);
 
-namespace Sylius\Bundle\MailerBundle\tests\Functional;
+namespace Sylius\Bundle\MailerBundle\Tests\Functional;
 
-use Sylius\Bundle\MailerBundle\tests\Provider\MessagesProvider;
-use Sylius\Bundle\MailerBundle\tests\Purger\SentMessagesPurger;
+use Sylius\Bundle\MailerBundle\Tests\Provider\MessagesProvider;
+use Sylius\Bundle\MailerBundle\Tests\Purger\SentMessagesPurger;
 use Sylius\Component\Mailer\Sender\SenderInterface;
 use Symfony\Bundle\FrameworkBundle\Test\KernelTestCase;
 use Symfony\Bundle\FrameworkBundle\Test\MailerAssertionsTrait;
